backend/requirements.txt
# Core dependencies
fastapi>=0.104
uvicorn[standard]
pydantic
python-dotenv
supabase
asyncpg
redis
PyJWT
python-jose
passlib[bcrypt]
httpx
python-multipart

# AI/ML dependencies
langchain
langgraph
openai
coqui-tts
faster-whisper
opencv-python
mediapipe

# D-ID Integration
requests>=2.31.0
websockets
python-socketio

# Audio Processing
sounddevice
numpy
scipy
librosa
